refactor(vo): Name 을 값 객체로 변경
